Extract gematria calculator to separate file
